refactor(browser): group protected browser modules
