+++ packages/platform-cloudflare/src/protected-browser.ts
@@ -1,2 +1,2 @@
-export { browserRunProtectedLayer } from "./protected-browser-policy.ts";
-export { browserRunProtectedBindingLayer } from "./protected-browser-native.ts";
+export { browserRunProtectedLayer } from "./protected-browser/policy.ts";
+export { browserRunProtectedBindingLayer } from "./protected-browser/binding.ts";
