+++ packages/platform-cloudflare/src/protected-browser/policy.ts
